fix(MapLegend): aria color contrast
